fix(doctor): satisfy migration pipeline guards
