+++ src/plugin-sdk/runtime-doctor-migrations.ts
@@ -6,13 +6,10 @@
  * uninstall flows). Those stay on `runtime-doctor`.
  */
 import fs from "node:fs/promises";
-import { buildLegacyMigrationPreview } from "../channels/plugins/legacy-state-migration-preview.js";
-import type { ChannelLegacyStateMigrationPlan } from "../channels/plugins/legacy-state-migration.types.js";
 import { asObjectRecord } from "../config/channel-compat-normalization.js";
 import type { CompatMutationResult } from "../config/channel-compat-normalization.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
 import type { OpenKeyedStoreOptions } from "../plugin-state/plugin-state-store.js";
-import type { PluginDoctorStateMigration } from "../plugins/doctor-contract-registry.js";
 import { archiveLegacyStateSource } from "../plugins/doctor-state-migration-fs.js";
 
 export { collectProviderDangerousNameMatchingScopes } from "../config/dangerous-name-matching.js";
@@ -50,12 +47,13 @@ export type {
 export type {
   PluginDoctorStateMigration,
   PluginDoctorStateMigrationContext,
-} from "../plugins/doctor-contract-registry.js";
+} from "../plugins/doctor-contract-module.js";
 export {
   archiveLegacyStateSource,
   legacyStateFileExists,
 } from "../plugins/doctor-state-migration-fs.js";
 export { buildLegacyMigrationPreview } from "../channels/plugins/legacy-state-migration-preview.js";
+export { definePluginDoctorMigrationFromPlans } from "../plugins/doctor-state-migration-plan-adapter.js";
 export type { DoctorSessionRouteStateOwner } from "../plugins/doctor-session-route-state-owner-types.js";
 
 type KeyMoveValue = { value: unknown };
@@ -66,60 +64,6 @@ type KeyMoveChangeContext = {
   targetValue: unknown;
   mappedValue: unknown;
 };
-
-/** Adapts legacy channel migration plans to the canonical plugin doctor contract. */
-export function definePluginDoctorMigrationFromPlans(params: {
-  id: string;
-  label: string;
-  doctorOnly?: boolean;
-  resolvePlans: (params: {
-    cfg: OpenClawConfig;
-    env: NodeJS.ProcessEnv;
-    stateDir: string;
-    oauthDir: string;
-  }) =>
-    | ChannelLegacyStateMigrationPlan[]
-    | Promise<ChannelLegacyStateMigrationPlan[] | null | undefined>
-    | null
-    | undefined;
-}): PluginDoctorStateMigration {
-  const resolvePlans = async (input: {
-    config: OpenClawConfig;
-    env: NodeJS.ProcessEnv;
-    stateDir: string;
-    oauthDir: string;
-  }): Promise<ChannelLegacyStateMigrationPlan[]> => {
-    const plans =
-      (await params.resolvePlans({
-        cfg: input.config,
-        env: input.env,
-        stateDir: input.stateDir,
-        oauthDir: input.oauthDir,
-      })) ?? [];
-    return plans.map((plan) =>
-      plan.kind === "plugin-state-import" && !plan.stateDir
-        ? { ...plan, stateDir: input.stateDir }
-        : plan,
-    );
-  };
-
-  return {
-    id: params.id,
-    label: params.label,
-    ...(params.doctorOnly === true ? { doctorOnly: true } : {}),
-    async detectLegacyState(input) {
-      const plans = await resolvePlans(input);
-      return plans.length > 0
-        ? { preview: plans.map((plan) => buildLegacyMigrationPreview(plan)) }
-        : null;
-    },
-    async migrateLegacyState(input) {
-      const plans = await resolvePlans(input);
-      const { runLegacyMigrationPlans } = await import("../infra/state-migrations.plugin-state.js");
-      return await runLegacyMigrationPlans(plans);
-    },
-  };
-}
 
 /** Collects a channel's root config and object-shaped account overrides in config order. */
 export function collectChannelAccountScopes(params: {
